Начало переноса из контролера в компоненты методов
