+++ src/components/task-edit.js
@@ -1,6 +1,8 @@
 import {AbstractComponent} from './abstract-component';
 
 import {Colors} from './data';
+import {render, unrender} from "./utils";
+import {Deadline} from "./deadline";
 
 export class TaskEdit extends AbstractComponent {
   constructor({description, tags, color, dueDate, repeatingDays, isFavorite, isArchive}, index) {
@@ -13,8 +15,8 @@ export class TaskEdit extends AbstractComponent {
     this._isFavorite = isFavorite;
     this._isArchive = isArchive;
     this._id = index;
+    this.toggleDate = this.toggleDate.bind(this);
   }
-
 
   _repeatingDaysCheck() {
     let result = false;
@@ -28,6 +30,45 @@ export class TaskEdit extends AbstractComponent {
     }
     return result;
   }
+
+  toggleDate() {
+    this._taskEdit.getElement()
+      .querySelector(`.card__date-deadline-toggle`)
+      .addEventListener(`click`, () => {
+        this._dateStatus = this._taskEdit.getElement().querySelector(`.card__date-status`);
+        switch (this._dateStatus.textContent.toLowerCase().trim()) {
+          case `yes`:
+            this._dateStatus.textContent = `no`;
+            const deadline = this._taskEdit.getElement()
+              .querySelector(`.card__date-deadline`);
+            deadline.querySelector(`.card__date`).value = ``;
+            unrender(deadline);
+            this._tasks.dueDate = ``;
+            break;
+          case `no`:
+            this._dateStatus.textContent = `yes`;
+            this._deadline = new Deadline(this._tasks);
+            const elementBeforeDeadline = this._taskEdit.getElement()
+              .querySelector(`.card__date-deadline-toggle`);
+            render(elementBeforeDeadline, this._deadline.getTemplate(), `afterend`);
+            break;
+        }
+      });
+  }
+
+  // _getMiniTemplate() {
+  //   return `<fieldset class="card__date-deadline">
+  //             <label class="card__input-deadline-wrap">
+  //               <input
+  //                 class="card__date"
+  //                 type="text"
+  //                 placeholder=""
+  //                 name="date"
+  //                 value="${new Date(this._dueDate)}"
+  //               />
+  //             </label>
+  //           </fieldset>`;
+  // }
 
   getTemplate() {
     return `<article class="card card--edit card--${this._color} ${Object.keys(this._repeatingDays).some((day) => this._repeatingDays[day]) ? `card--repeat` : ``}">
@@ -63,6 +104,8 @@ export class TaskEdit extends AbstractComponent {
       <div class="card__settings">
         <div class="card__details">
           <div class="card__dates">
+
+
 ${this._dueDate === null ?
     `<button class="card__date-deadline-toggle" type="button">
               date: <span class="card__date-status">no</span>
@@ -85,7 +128,7 @@ ${this._dueDate === null ?
             <button class="card__repeat-toggle" type="button">
               repeat:<span class="card__repeat-status">${this._repeatingDaysCheck() ? `yes` : `no`}</span>
             </button>
-            
+
             <fieldset class="card__repeat-days">
               <div class="card__repeat-days-inner">
               ${this._repeatingDaysCheck() ? Object.keys(this._repeatingDays).map((day) =>`<input
@@ -164,4 +207,5 @@ ${this._dueDate === null ?
   </form>
 </article>`;
   }
+
 }
